Fix string ... to int... b....

+++ python/heis/elevator_system.py
@@ -88,7 +88,7 @@ class System:
             if self.elevators[elevator]['current_floor'] == None:
                 current_elevator = elevator
             else:
-                if abs(int(self.elevators[elevator]['current_floor']) - floor) < abs(int(self.elevators[current_elevator]['current_floor']) - floor):
+                if abs(int(self.elevators[elevator]['current_floor']) - int(floor)) < abs(int(self.elevators[current_elevator]['current_floor']) - int(floor)):
                     if self.elevators[elevator]['direction'] == self.elevators[current_elevator]['direction'] == direction:
                         current_elevator = elevator
